Generate pkm qr better
Don't include nature/ability for pre-gen3, export held item string with
correct version context
(whatever for cxd)

Closes #1915
